+++ core/src/main/kotlin/com/gitlab/kordlib/core/live/channel/LiveNewsChannel.kt
@@ -10,9 +10,12 @@ import com.gitlab.kordlib.core.event.channel.NewsChannelUpdateEvent
 import com.gitlab.kordlib.core.event.guild.GuildDeleteEvent
 
 @KordPreview
+@Suppress("DeprecatedCallableAddReplaceWith")
+@Deprecated("Message channels are allowed to change type during their lifetime. As such, there's no guarantee that the channel will stay a NewsChannel", level = DeprecationLevel.WARNING)
 fun NewsChannel.live() = LiveNewsChannel(this)
 
 @KordPreview
+@Deprecated("Message channels are allowed to change type during their lifetime. As such, there's no guarantee that the channel will stay a NewsChannel", level = DeprecationLevel.WARNING)
 class LiveNewsChannel(channel: NewsChannel) : LiveChannel(), Entity by channel {
 
     override var channel: NewsChannel = channel