+++ core/src/main/kotlin/com/gitlab/kordlib/core/live/channel/LiveStoreChannel.kt
@@ -11,9 +11,12 @@ import com.gitlab.kordlib.core.event.channel.StoreChannelUpdateEvent
 import com.gitlab.kordlib.core.event.guild.GuildDeleteEvent
 
 @KordPreview
+@Suppress("DeprecatedCallableAddReplaceWith")
+@Deprecated("Message channels are allowed to change type during their lifetime. As such, there's no guarantee that the channel will stay a StoreChannel", level = DeprecationLevel.WARNING)
 fun StoreChannel.live() = LiveStoreChannel(this)
 
 @KordPreview
+@Deprecated("Message channels are allowed to change type during their lifetime. As such, there's no guarantee that the channel will stay a StoreChannel", level = DeprecationLevel.WARNING)
 class LiveStoreChannel(channel: StoreChannel) : LiveChannel(), Entity by channel {
 
     override var channel: StoreChannel = channel