+++ core/src/main/kotlin/com/gitlab/kordlib/core/live/channel/LiveTextChannel.kt
@@ -10,9 +10,12 @@ import com.gitlab.kordlib.core.event.channel.TextChannelUpdateEvent
 import com.gitlab.kordlib.core.event.guild.GuildDeleteEvent
 
 @KordPreview
+@Suppress("DeprecatedCallableAddReplaceWith")
+@Deprecated("Message channels are allowed to change type during their lifetime. As such, there's no guarantee that the channel will stay a TextChannel", level = DeprecationLevel.WARNING)
 fun TextChannel.live() = LiveTextChannel(this)
 
 @KordPreview
+@Deprecated("Message channels are allowed to change type during their lifetime. As such, there's no guarantee that the channel will stay a TextChannel", level = DeprecationLevel.WARNING)
 class LiveTextChannel(channel: TextChannel) : LiveChannel(), Entity by channel {
 
     override var channel: TextChannel = channel